+++ components/org.wso2.carbon.micro.integrator.core/src/main/java/org/wso2/carbon/core/util/KeyStoreManager.java
@@ -19,10 +19,9 @@ package org.wso2.carbon.core.util;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.core.util.CarbonException;
 import org.wso2.carbon.core.RegistryResources;
 import org.wso2.carbon.micro.integrator.core.internal.CarbonCoreDataHolder;
-import org.wso2.carbon.micro.integrator.core.internal.CarbonServerConfigurationService;
+import org.wso2.carbon.micro.integrator.core.services.CarbonServerConfigurationService;
 import org.wso2.carbon.registry.core.Resource;
 import org.wso2.carbon.registry.core.service.RegistryService;
 import org.wso2.carbon.micro.integrator.core.util.MicroIntegratorBaseUtils;
