+++ components/org.wso2.carbon.micro.integrator.core/src/main/java/org/wso2/carbon/micro/integrator/core/services/Axis2ConfigurationContextService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.carbon.micro.integrator.core.internal;
+package org.wso2.carbon.micro.integrator.core.services;
 
 import org.apache.axis2.context.ConfigurationContext;
